pyproject.toml
[tool.poetry]
name = "oia-correct-rate-pipeline"
version = "0.1.0"
description = ""
authors = ["momonong <morrischen2.9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
gspread = "^6.2.0"
google-auth = "^2.38.0"
openai = "^1.68.2"
pandas = "^2.2.3"
python-dotenv = "^1.1.0"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
